feat: Add new weight generator
