+++ SNN40nm.py
@@ -17,7 +17,7 @@ from loguru import logger
 from BrainpyAdapter.BrainpyBase import BrainpyBase
 from Common.Common import SpikeWriter, convert_spike_npy_to_bin, div_round_up
 from Mapping.Router import Router40nm
-from Mapping.SNNweight import SNNWeight
+from Mapping.Weight import Weight40nm
 from SNNCompiler.hardware_config import HardwareConfig
 
 
@@ -52,7 +52,8 @@ class SNN40nm():
         t1 = time.time()
         self.debug.record_running_time(t1-t0, label='Network analysis')
 
-        self.used_tile_num = int(np.ceil(self.neuron_num/(16*1024)))
+        self.used_tile_num = int(np.ceil(
+            self.neuron_num/(self.config['Tile_NpuNum']*self.config['Npu_NeuronNum'])))
         self.used_tile_cols = div_round_up(
             self.used_tile_num, self.config['Y_TileNum'])
         self.used_tile_rows = self.used_tile_num if self.used_tile_cols == 1 else self.config[
@@ -65,10 +66,11 @@ class SNN40nm():
 
         # weight data dump
         t0 = time.time()
-        network = SNNWeight(
-            self.connection_matrix, self.neuron_num, self.neuron_scale, self.config)
-        network.dump(download_dir)
+        weight = Weight40nm(self.connection_matrix,
+                            self.neuron_num, self.neuron_scale, self.config)
+        weight.write_to_bin(download_dir)
         t1 = time.time()
+
         self.debug.record_running_time(t1-t0, label='Weight bin data')
 
         # spike data dump
